+++ lib/avromatic/model/builder.rb
@@ -4,6 +4,7 @@ require 'active_model'
 require 'avromatic/model/configuration'
 require 'avromatic/model/value_object'
 require 'avromatic/model/configurable'
+require 'avromatic/model/nested_models'
 require 'avromatic/model/attribute/union'
 require 'avromatic/model/attributes'
 require 'avromatic/model/attribute/record'
@@ -42,6 +43,7 @@ module Avromatic
           ActiveModel::Validations,
           Virtus.value_object,
           Avromatic::Model::Configurable,
+          Avromatic::Model::NestedModels,
           Avromatic::Model::Attributes,
           Avromatic::Model::ValueObject,
           Avromatic::Model::RawSerialization,