+++ lib/avromatic/model/value_object.rb
@@ -1,5 +1,6 @@
 module Avromatic
   module Model
+    extend ActiveSupport::Concern
 
     # This module is used to override the comparisons defined by
     # Virtus::Equalizer which is pulled in by Virtus::ValueObject.