+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v0.9.0 (unreleased)
 - Add support for more than one non-null type in a union.
+- Fix the serialization of nested complex types.
 
 ## v0.8.0
 - Add support for logical types. Currently this requires using the