+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # avromatic changelog
 
-## v0.9.0 (unreleased)
+## v0.9.0
 - Experimental: Add support for more than one non-null type in a union.
 - Allow nested models to be referenced and reused.
 - Fix the serialization of nested complex types.